Major changes. Read comments.
App is now responsive! Integrated bootstrap. Will need to integrate more as time goes on, but its good for now.

In order to do this, buttons and player info were removed in favor of a dropdown menu because it looks nicer on mobile.

Also I was able to remove a lot of now unnecessary CSS.

Refactored some and moved more util functions into utils.js